requirements.txt
boto3==1.28.84
botocore==1.31.84
python-dotenv==1.0.0
